Merge: Service para Feature

+++ sistema-gerenciador-pedidos/src/main/java/com/tecdes/lanchonete/service/GerenteService.java
@@ -2,6 +2,8 @@ package com.tecdes.lanchonete.service;
 
 import java.util.List;
 
+import com.tecdes.lanchonete.exception.InvalidDeleteOperationException;
+import com.tecdes.lanchonete.exception.InvalidFieldException;
 import com.tecdes.lanchonete.generalinterfaces.crud.Crud;
 import com.tecdes.lanchonete.model.entity.Gerente;
 import com.tecdes.lanchonete.repository.implementation.IGerenteRepository;
@@ -19,12 +21,12 @@ public class GerenteService implements Crud<Gerente> {
     @Override
     public Gerente create(Gerente t) {
         if(t.getLogin() == null){
-
+            throw new InvalidFieldException("Login do gerente não pode ser nulo");
         }
         if(t.getSenha() == null){
-
+            throw new InvalidFieldException("Senha do gerente não pode ser nula");
         }
-
+        
         funcionarioService.create(t);
         return iGerenteRepository.create(t);
     }
@@ -41,12 +43,15 @@ public class GerenteService implements Crud<Gerente> {
 
     @Override
     public void update(Gerente t) {
-        funcionarioService.update(t);
+        funcionarioService.update(t); // Funcionário service já verifica se id é null
         iGerenteRepository.update(t);
     }
 
     @Override
     public void delete(Long id) {
+        if (id == null) {
+            throw new InvalidDeleteOperationException("Não é possível deletar Gerente nulo.");
+        }
         iGerenteRepository.delete(id);
         funcionarioService.delete(id);
     }
